+++ utils.h
@@ -8,10 +8,10 @@ These are some utility functions and variables used through the main code
 #define MS_IN_MINUTE 60000
 #define US_IN_MINUTE 60000000
 
-int freeRam ();
+short freeRam ();
 void setPinMapping(byte boardID);
-unsigned int PW();
-unsigned int PW_SD();
-unsigned int PW_AN();
+unsigned short PW();
+unsigned short PW_SD();
+unsigned short PW_AN();
 
 #endif // UTILS_H
